requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
jinja2==3.1.6
python-multipart==0.0.20
pydantic==2.11.7
itsdangerous==2.2.0
bcrypt==4.3.0
python-dotenv==1.1.0
